requirements.txt
google-generativeai
streamlit
pandas
openpyxl
pypdf

